refactor: rename data module to structs and update imports for PasswordEntry
