+++ src/models/mod.rs
@@ -1,2 +1,2 @@
-pub mod data;
+pub mod structs;
 pub mod types;
