+++ src/models/structs.rs
